Move e2e tests between screens
